completed watershed functionality
