+++ requirements.txt
@@ -1,17 +1,77 @@
-argparse==1.4.0
+astroid==3.0.1
+asttokens==2.4.1
+attrs==23.1.0
+certifi==2023.7.22
+cfgv==3.4.0
+click==8.1.7
+click-plugins==1.1.1
+cligj==0.7.2
+colorama==0.4.6
+comm==0.1.4
+cycler==0.12.1
+debugpy==1.8.0
+decorator==5.1.1
 descartes==1.1.0
+dill==0.3.7
+distlib==0.3.7
+et-xmlfile==1.1.0
+exceptiongroup==1.1.3
+executing==2.0.1
+filelock==3.13.1
+fiona==1.9.5
+fonttools==4.44.0
 geopandas==0.12.2
-ipykernel
+identify==2.5.31
+iniconfig==2.0.0
+ipykernel==6.26.0
+ipython==8.17.2
+isort==5.12.0
+jedi==0.19.1
+jupyter_client==8.5.0
+jupyter_core==5.5.0
+kiwisolver==1.4.5
 matplotlib==3.5.1
+matplotlib-inline==0.1.6
+mccabe==0.7.0
+memory-profiler==0.61.0
+mypy==1.6.1
+mypy-extensions==1.0.0
+nest-asyncio==1.5.8
 networkx==2.6.3
+nodeenv==1.8.0
 numpy==1.22.1
-openpyxl== 3.1.2
+openpyxl==3.1.2
+packaging==23.2
 pandas==1.4.0
-pre-commit
-pylint
-pytest
+parso==0.8.3
+Pillow==10.1.0
+platformdirs==3.11.0
+pluggy==1.3.0
+pre-commit==3.5.0
+prompt-toolkit==3.0.39
+psutil==5.9.6
+pure-eval==0.2.2
+Pygments==2.16.1
+pylint==3.0.2
+pyparsing==3.1.1
+pyproj==3.6.1
+pytest==7.4.3
 pytest-resource-usage==1.0.0
-ruff
+python-dateutil==2.8.2
+pytz==2023.3.post1
+pywin32==306
+PyYAML==6.0.1
+pyzmq==25.1.1
+ruff==0.1.4
 scipy==1.11.3
-Shapely==2.0.1
-toml
+shapely==2.0.1
+six==1.16.0
+stack-data==0.6.3
+toml==0.10.2
+tomli==2.0.1
+tomlkit==0.12.2
+tornado==6.3.3
+traitlets==5.13.0
+typing_extensions==4.8.0
+virtualenv==20.24.6
+wcwidth==0.2.9
